Moved definition of classification classes to avoid duplicates
