+++ tests/OTAnalytics/plugin_number_of_tracks_to_be_validated/test_calculation_strategy.py
@@ -1,6 +1,7 @@
 import pytest
 from pandas import DataFrame
 
+from OTAnalytics.adapter_visualization.otc_classes import OtcClasses
 from OTAnalytics.domain.track import TRACK_CLASSIFICATION, TRACK_ID, Track
 from OTAnalytics.domain.track_dataset import TRACK_GEOMETRY_FACTORY
 from OTAnalytics.plugin_datastore.track_store import PandasTrackDataset
@@ -10,7 +11,6 @@ from OTAnalytics.plugin_number_of_tracks_to_be_validated.calculation_strategy im
     DetectionRateByMaxConfidence,
     DetectionRateByPercentile,
 )
-from OTAnalytics.plugin_number_of_tracks_to_be_validated.otc_classes import OtcClasses
 from tests.OTAnalytics.plugin_number_of_tracks_to_be_validated.conftest import (
     TRACK_ONE_FALSE_DETECTION_CLASS,
     TRACK_PERFECT,
